refactor!: Move tableprinter into internal package
