+++ internal/tableprinter/table_printer_test.go
@@ -21,7 +21,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package utils
+package tableprinter
 
 import (
 	"bytes"
